added all the icons for organisms
