--- src/grass.js
@@ -1,7 +1,0 @@
-import {Plant} from './plant.js';
-
-export class Grass extends Plant {
-    constructor() {
-        super();
-    }
-}
