--- src/guarana.js
@@ -1,7 +1,0 @@
-import {Plant} from './plant.js';
-
-export class Guarana extends Plant {
-    constructor() {
-        super();
-    }
-}
