+++ src/organismsList.js
@@ -1,12 +1,12 @@
-import {Sheep} from "./sheep";
-import {Wolf} from "./wolf";
-import {Fox} from "./fox/Fox";
+import {Sheep} from "./sheep/Sheep.js";
+import {Wolf} from "./wolf/Wolf.js";
+import {Fox} from "./fox/Fox.js";
 import {Antelope} from "./antelope/Antelope.js";
-import {Turtle} from "./turtle";
-import {Grass} from "./grass";
-import {Guarana} from "./guarana";
-import {PoisonBerry} from "./poisonBerry";
-import {SowThistle} from "./snowThistle";
+import {Turtle} from "./turtle/Turtle.js";
+import {Grass} from "./grass/Grass.js";
+import {Guarana} from "./guarana/Guarana.js";
+import {Berry} from "./berry/Berry.js";
+import {SowThistle} from "./thistle/Thistle.js";
 
 
-export let organismsList = [Wolf, Sheep, Fox, Antelope, Turtle, Grass, Guarana, PoisonBerry, SowThistle];
+export let organismsList = [Wolf, Sheep, Fox, Antelope, Turtle, Grass, Guarana, Berry, SowThistle];
