--- src/poisonBerry.js
@@ -1,7 +1,0 @@
-import {Plant} from './plant.js';
-
-export class PoisonBerry extends Plant {
-    constructor() {
-        super();
-    }
-}
