--- src/sheep.js
@@ -1,9 +1,0 @@
-import {Animal} from './animal.js';
-
-export class Sheep extends Animal {
-    constructor() {
-        super();
-        this.initiative = 4;
-        this.strenght = 3;
-    }
-}
