--- src/snowThistle.js
@@ -1,7 +1,0 @@
-import {Plant} from './plant.js';
-
-export class SowThistle extends Plant {
-    constructor() {
-        super();
-    }
-}
