--- src/turtle.js
@@ -1,9 +1,0 @@
-import {Animal} from './animal.js';
-
-export class Turtle extends Animal {
-    constructor() {
-        super();
-        this.initiative = 1;
-        this.strenght = 2;
-    }
-}
