--- src/wolf.js
@@ -1,10 +1,0 @@
-import {Animal} from './animal.js';
-
-export class Wolf extends Animal {
-    constructor() {
-        super();
-        this.initiative = 5;
-        this.strenght = 9;
-
-    }
-}
