+++ src/wolf/Wolf.js
@@ -1,0 +1,16 @@
+import {Animal} from '../animal.js';
+import wolfImage from "./wolf.jpg";
+
+export class Wolf extends Animal {
+    constructor() {
+        super();
+        this.initiative = 5;
+        this.strenght = 9;
+        this.createElement()
+    }
+
+    createElement() {
+        super.createElement()
+        this.element.src = wolfImage;
+    }
+}
